# artifacts/api-server/src/routes/playback.ts
import { Router } from "express";
import { db, devicesTable, clientsTable, playlistsTable, playlistItemsTable, mediaTable, playbackLogsTable } from "@workspace/db";
import { eq, and, asc } from "drizzle-orm";

const router = Router();

/** Shared helper: verify device is active and return device + client */
async function resolveDevice(uuid: string, email: string) {
  const [device] = await db
    .select()
    .from(devicesTable)
    .where(and(eq(devicesTable.uuid, uuid), eq(devicesTable.email, email)))
    .limit(1);

  if (!device || device.status !== "active") return { error: "Forbidden", message: "Device not authorized" };
  if (!device.clientId) return { error: "Forbidden", message: "Device not associated with a client" };

  const [client] = await db.select().from(clientsTable).where(eq(clientsTable.id, device.clientId)).limit(1);
  if (!client) return { error: "Not Found", message: "Client not found" };

  return { device, client };
}

// Public - list all active playlists for the device's client
router.get("/playback/playlists", async (req, res) => {
  const { uuid, email } = req.query;
  if (!uuid || !email) {
    res.status(400).json({ error: "Bad Request", message: "uuid and email required" });
    return;
  }

  const result = await resolveDevice(uuid as string, email as string);
  if ("error" in result) {
    res.status(result.error === "Not Found" ? 404 : 403).json(result);
    return;
  }

  const { client } = result;

  const playlists = await db
    .select({
      id: playlistsTable.id,
      name: playlistsTable.name,
    })
    .from(playlistsTable)
    .where(and(eq(playlistsTable.clientId, client.id), eq(playlistsTable.active, true)))
    .orderBy(asc(playlistsTable.id));

  // Get item counts per playlist
  const playlistsWithCount = await Promise.all(
    playlists.map(async (pl) => {
      const items = await db
        .select({ id: playlistItemsTable.id })
        .from(playlistItemsTable)
        .where(eq(playlistItemsTable.playlistId, pl.id));
      return { id: pl.id, name: pl.name, itemCount: items.length };
    })
  );

  res.json(playlistsWithCount);
});

// Public - get playback queue for a device
router.get("/playback/queue", async (req, res) => {
  const { uuid, email, playlistId } = req.query;
  if (!uuid || !email) {
    res.status(400).json({ error: "Bad Request", message: "uuid and email required" });
    return;
  }

  const result = await resolveDevice(uuid as string, email as string);
  if ("error" in result) {
    res.status(result.error === "Not Found" ? 404 : 403).json(result);
    return;
  }

  const { device, client } = result;

  // Try to load the requested playlist; fall back to first active if omitted/invalid
  let playlist: { id: number; name: string; clientId: number; playbackMode: string; active: boolean } | undefined;

  if (playlistId) {
    const requestedId = parseInt(playlistId as string, 10);
    if (!isNaN(requestedId)) {
      // Only allow playlist that belongs to this client
      const [found] = await db
        .select()
        .from(playlistsTable)
        .where(and(eq(playlistsTable.id, requestedId), eq(playlistsTable.clientId, client.id), eq(playlistsTable.active, true)))
        .limit(1);
      playlist = found;
    }
  }

  if (!playlist) {
    // Fallback: first active playlist for the client
    const [first] = await db
      .select()
      .from(playlistsTable)
      .where(and(eq(playlistsTable.clientId, client.id), eq(playlistsTable.active, true)))
      .orderBy(asc(playlistsTable.id))
      .limit(1);
    playlist = first;
  }

  if (!playlist) {
    res.json({
      clientId: client.id,
      deviceId: device.id,
      playlistId: null,
      currentIndex: 0,
      playbackMode: client.playbackMode,
      jingleMode: client.jingleMode,
      jingleInterval: client.jingleInterval,
      jingleIntervalSeconds: client.jingleIntervalSeconds,
      musicVolume: 1.0,
      jingleVolume: 0.8,
      items: [],
    });
    return;
  }

  const items = await db
    .select({
      id: mediaTable.id,
      title: mediaTable.title,
      artist: mediaTable.artist,
      type: mediaTable.type,
      filename: mediaTable.filename,
      filePath: mediaTable.filePath,
      duration: mediaTable.duration,
      clientId: mediaTable.clientId,
      gain: mediaTable.gain,
      coverUrl: mediaTable.coverUrl,
      createdAt: mediaTable.createdAt,
    })
    .from(playlistItemsTable)
    .innerJoin(mediaTable, eq(playlistItemsTable.mediaId, mediaTable.id))
    .where(eq(playlistItemsTable.playlistId, playlist.id))
    .orderBy(asc(playlistItemsTable.position));

  const itemsWithUrl = items.map((item) => ({
    ...item,
    url: `/api/uploads/${item.filename}`,
  }));

  res.json({
    clientId: client.id,
    deviceId: device.id,
    playlistId: playlist.id,
    currentIndex: 0,
    playbackMode: client.playbackMode,
    jingleMode: client.jingleMode,
    jingleInterval: client.jingleInterval,
    jingleIntervalSeconds: client.jingleIntervalSeconds,
    musicVolume: 1.0,
    jingleVolume: 0.8,
    items: itemsWithUrl,
  });
});

// Public - log a playback event
router.post("/playback/log", async (req, res) => {
  const { mediaId, uuid, email, playedAt } = req.body;
  if (!mediaId || !uuid || !email) {
    res.status(400).json({ error: "Bad Request", message: "mediaId, uuid and email required" });
    return;
  }

  const [device] = await db.select().from(devicesTable).where(eq(devicesTable.uuid, uuid)).limit(1);

  await db.insert(playbackLogsTable).values({
    mediaId,
    deviceId: device?.id ?? null,
    deviceUuid: uuid,
    clientEmail: email,
    playedAt: playedAt ? new Date(playedAt) : new Date(),
  });

  res.status(201).json({ success: true });
});

export default router;