# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Indoor Radio System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List media files
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [music, jingle]
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaItem"
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a media file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadMediaBody"
      responses:
        "201":
          description: Media uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
  /media/scan-folder:
    post:
      operationId: scanFolder
      tags: [media]
      summary: Scan a server folder and register found audio files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanFolderBody"
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanFolderResult"
  /media/batch-delete:
    post:
      operationId: deleteMediaBatch
      tags: [media]
      summary: Delete multiple media files at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mediaIds:
                  type: array
                  items:
                    type: integer
              required: [mediaIds]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                required: [deleted]
  /media/{mediaId}:
    get:
      operationId: getMedia
      tags: [media]
      summary: Get media metadata
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media file
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /playback/playlists:
    get:
      operationId: getPlaybackPlaylists
      tags: [playback]
      summary: List all active playlists available to the device's client
      parameters:
        - name: uuid
          in: query
          required: true
          schema:
            type: string
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistSummary"
        "403":
          description: Device not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /playback/queue:
    get:
      operationId: getPlaybackQueue
      tags: [playback]
      summary: Get the playback queue for the current device
      parameters:
        - name: uuid
          in: query
          required: true
          schema:
            type: string
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: playlistId
          in: query
          required: false
          schema:
            type: integer
          description: Optional playlist ID to load. Falls back to the first active playlist if omitted or invalid.
      responses:
        "200":
          description: Playback queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackQueue"
        "403":
          description: Device not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /playback/log:
    post:
      operationId: logPlayback
      tags: [playback]
      summary: Log a completed media playback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogPlaybackBody"
      responses:
        "201":
          description: Log recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    MediaItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        type:
          type: string
          enum: [music, jingle]
        filename:
          type: string
        url:
          type: string
        duration:
          type: number
          description: Duration in seconds
        clientId:
          type: integer
        gain:
          type: number
          description: Volume gain preset (0.0 to 2.0)
        coverUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - filename
        - url
        - clientId
        - createdAt
    UploadMediaBody:
      type: object
      properties:
        file:
          type: string
          format: binary
        title:
          type: string
        artist:
          type: string
        type:
          type: string
          enum: [music, jingle]
        clientId:
          type: integer
        gain:
          type: number
      required:
        - file
        - title
        - type
        - clientId
    ScanFolderBody:
      type: object
      properties:
        folderPath:
          type: string
        clientId:
          type: integer
        type:
          type: string
          enum: [music, jingle]
      required:
        - folderPath
        - clientId
        - type
    ScanFolderResult:
      type: object
      properties:
        found:
          type: integer
        imported:
          type: integer
        skipped:
          type: integer
        files:
          type: array
          items:
            type: string
      required:
        - found
        - imported
        - skipped
        - files
...
    PlaylistSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        itemCount:
          type: integer
      required:
        - id
        - name
        - itemCount
    PlaybackQueue:
      type: object
      properties:
        clientId:
          type: integer
        deviceId:
          type: integer
        playlistId:
          type: integer
          description: The ID of the playlist whose items are in this queue
        currentIndex:
          type: integer
        playbackMode:
          type: string
          enum: [sequential, shuffle]
        jingleMode:
          type: string
          enum: [ordered, interval, time]
        jingleInterval:
          type: integer
        jingleIntervalSeconds:
          type: integer
        musicVolume:
          type: number
        jingleVolume:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
      required:
        - clientId
        - deviceId
        - currentIndex
        - playbackMode
        - jingleMode
        - musicVolume
        - jingleVolume
        - items
    LogPlaybackBody:
      type: object
      properties:
        mediaId:
          type: integer
        uuid:
          type: string
        email:
          type: string
        playedAt:
          type: string
          format: date-time
      required:
        - mediaId
        - uuid
        - email
...